+++ CHANGES.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Improved cache configuration when using plugins (usage: --connector kvm:::cache=true,cache_size=1kb,cache_time=10,cache_page_size=1000 where size and page_size is specified in hex)
 - Added DelayedPhysicalMemory middleware (usage: --connector kvm:::delay=200 where delay is specified in microseconds)
 - Added PhysicalMemoryMetrics middleware (usage: --connector kvm:::metrics=true)
+- Updated FileIoMemory constructor with a default identity mapped memory mapping.
 
 ## 0.2.0-beta9
 ## 0.2.0-beta8